requirements.txt
# ============================================================
# Neural Automater — Requirements
# Install: pip install -r requirements.txt
# Then: playwright install chromium
# ============================================================

# Core GUI
customtkinter
Pillow
pyautogui

# Browser Automation
playwright

# AI / LLM Providers
google-generativeai
openai

# Voice
SpeechRecognition
pyaudio          # Windows: pipwin install pyaudio
pyttsx3
pvporcupine

# Web & HTML
requests
beautifulsoup4
duckduckgo-search

# Data
pandas
numpy
schedule

# Email & Calendar
google-auth
google-auth-oauthlib
google-api-python-client

# Phone / SMS
twilio

# Trading
python-binance

# Memory & RAG (V2)
chromadb
PyPDF2
python-docx
sentence-transformers

# File Auth
cryptography

# Utilities
python-dotenv

# Business Scraper
openpyxl
